+++ src/plugins/browser/additional-notifiers.js
@@ -1,6 +1,5 @@
-
 function plugin(loggin) {
-    const { Notifier, Pipe } = loggin;
+    const { Notifier } = loggin;
 
     class ConsoleNotifier extends Notifier {
         constructor(options) {
@@ -21,7 +20,26 @@ function plugin(loggin) {
         }
     }
 
+    class HttpNotifier extends Notifier {
+        constructor(options) {
+            super(options, 'http');
+            this.headers = this.options.headers || {};
+        }
+
+        async output(logMsg, log) {
+            return await fetch(this.options.url, {
+                method: 'POST',
+                headers: this.headers,
+                body: JSON.stringify({
+                    message: logMsg,
+                    log
+                })
+            });
+        }
+    }
+
     Notifier.register('Console', ConsoleNotifier);
+    Notifier.register('Http', HttpNotifier);
 };
 
 module.exports = plugin;